+++ applications/content/src/main/java/org/apache/ofbiz/content/content/ContentKeywordIndex.java
@@ -70,7 +70,7 @@ public class ContentKeywordIndex {
         List<String> strings = new LinkedList<String>();
 
         int pidWeight = 1;
-        keywords.put(content.getString("contentId").toLowerCase(Locale.getDefault()), Long.valueOf(pidWeight));
+        keywords.put(content.getString("contentId").toLowerCase(Locale.getDefault()), (long) pidWeight);
 
         addWeightedKeywordSourceString(content, "dataResourceId", strings);
         addWeightedKeywordSourceString(content, "contentName", strings);
@@ -187,7 +187,7 @@ public class ContentKeywordIndex {
         }
 
         List<GenericValue> toBeStored = new LinkedList<GenericValue>();
-        int keywordMaxLength = EntityUtilProperties.getPropertyAsInteger("contentsearch", "content.keyword.max.length", 0).intValue();
+        int keywordMaxLength = EntityUtilProperties.getPropertyAsInteger("contentsearch", "content.keyword.max.length", 0);
         for (Map.Entry<String, Long> entry: keywords.entrySet()) {
             if (entry.getKey().length() <= keywordMaxLength) {
                 GenericValue contentKeyword = delegator.makeValue("ContentKeyword", UtilMisc.toMap("contentId", content.getString("contentId"), "keyword", entry.getKey(), "relevancyWeight", entry.getValue()));
