+++ applications/product/src/main/java/org/apache/ofbiz/product/category/ftl/UrlRegexpTransform.java
@@ -221,7 +221,7 @@ public class UrlRegexpTransform implements TemplateTransformModel {
             if (responseCodeInt == null) {
                 response.setStatus(SeoConfigUtil.getDefaultResponseCode());
             } else {
-                response.setStatus(responseCodeInt.intValue());
+                response.setStatus(responseCodeInt);
             }
             response.setHeader("Location", uri);
         } else {
